Split player code into folders for base player and app
This also extract granular interfaces for `class Environment`.

This aims to make the dependency structure more explicit, and to make
the base player easier to reuse.
