+++ Resources/Private/JavaScript/VideoPlayer/controls/FlatSeekBar.js
@@ -2,10 +2,10 @@
 
 import shaka from 'shaka-player/dist/shaka-player.ui';
 
+import { e } from '../../lib/util';
 import Chapters from '../Chapters';
 import ImageFetcher from '../ImageFetcher';
 import ThumbnailPreview from '../ThumbnailPreview';
-import { e } from '../util';
 import VariantGroups from '../VariantGroups';
 
 /**
