+++ Resources/Private/JavaScript/VideoPlayer/controls/index.js
@@ -1,0 +1,7 @@
+// @ts-check
+
+export { default as ControlPanelButton } from './ControlPanelButton';
+export { default as FlatSeekBar } from './FlatSeekBar';
+export { default as OverflowMenuButton } from './OverflowMenuButton';
+export { default as PresentationTimeTracker } from './PresentationTimeTracker';
+export { default as VideoTrackSelection } from './VideoTrackSelection';
