+++ Resources/Private/JavaScript/VideoPlayer/index.js
@@ -1,0 +1,6 @@
+// @ts-check
+
+export { default as Chapters } from './Chapters';
+export { ControlPanelButton, OverflowMenuButton } from './controls';
+export { default as buildTimeString } from './lib/buildTimeString';
+export { default as SachsenShakaPlayer } from './SachsenShakaPlayer';
