+++ Resources/Private/JavaScript/VideoPlayerApp/Screenshot.js
