+++ Resources/Private/JavaScript/VideoPlayerApp/index.js
@@ -1,0 +1,5 @@
+// @ts-check
+
+import SxndPlayerApp from './SxndPlayerApp';
+
+window.SxndPlayerApp = SxndPlayerApp;
