+++ Resources/Private/JavaScript/VideoPlayerApp/keybindings.json
