+++ Resources/Private/JavaScript/VideoPlayerApp/lib/Component.js
