+++ Resources/Private/JavaScript/VideoPlayerApp/lib/Modals.js
