+++ Resources/Private/JavaScript/VideoPlayerApp/lib/SimpleModal.js
@@ -2,8 +2,8 @@
 
 import $ from 'jquery';
 
+import { e } from '../../lib/util';
 import Component from './Component';
-import { e } from './util';
 
 /**
  * @typedef {{
