+++ Resources/Private/JavaScript/VideoPlayerApp/modals/index.js
@@ -1,0 +1,5 @@
+// @ts-check
+
+export { default as BookmarkModal } from './BookmarkModal';
+export { default as HelpModal } from './HelpModal';
+export { default as ScreenshotModal } from './ScreenshotModal';
