+++ Resources/Private/JavaScript/lib/image/createImage.js
