+++ Resources/Private/JavaScript/lib/image/imageFormats.js
