+++ Resources/Private/JavaScript/lib/image/jpeg/index.js
