+++ Resources/Private/JavaScript/lib/image/jpeg/index.test.js
