+++ Resources/Private/JavaScript/lib/image/png/index.js
