+++ Resources/Private/JavaScript/lib/image/png/simple.png
