+++ Resources/Private/JavaScript/lib/util.js
@@ -1,18 +1,6 @@
 // @ts-check
 
 /**
- *
- * @param {MetadataArray} metadataArray
- * @returns {string}
- */
-export function metadataArrayToString(metadataArray) {
-  return metadataArray.screenshotFields
-    // TODO: Find a better way
-    .map(field => metadataArray.metadata[field]?.[0] ?? "")
-    .join(' / ');
-}
-
-/**
  * Clamps {@link value} into the closed interval [{@link min}, {@link max}].
  *
  * @param {number} value
@@ -29,37 +17,6 @@ export function clamp(value, [min, max]) {
   }
 
   return value;
-}
-
-/**
- * Formats {@link totalSeconds} to a time string.
- *
- * The base format is `hh:mm:ss:ff`. Hours and frames are included depending on
- * {@link showHour} and {@link fps}. The first part is not zero-padded.
- *
- * Adopted from shaka.ui.Utils.buildTimeString.
- *
- * @param {number} totalSeconds Total number of seconds to be formatted.
- * @param {boolean} showHour Whether or not to show hours.
- * @param {number | null} fps (Optional) Number of FPS used to calculate frame
- * count.
- * @returns {string}
- */
-export function buildTimeString(totalSeconds, showHour, fps = null) {
-  const segments = showHour
-    ? [totalSeconds / 3600, (totalSeconds / 60) % 60, totalSeconds % 60]
-    : [totalSeconds / 60, totalSeconds % 60];
-
-  if (fps) {
-    segments.push(Math.floor((totalSeconds % 1) * fps));
-  }
-
-  return (
-    segments
-      // Don't pad the first segment
-      .map((n, i) => Math.floor(n).toString().padStart(i == 0 ? 0 : 2, '0'))
-      .join(':')
-  );
 }
 
 /**
@@ -224,7 +181,7 @@ export function binaryStringToArrayBuffer(s) {
  *
  * @template {keyof HTMLElementTagNameMap} K
  * @param {K} tag
- * @param {Partial<ElementAttributes<K>> & Partial<EventListeners>} attrs
+ * @param {Partial<HTMLElementTagNameMap[K]> & Partial<EventListeners<'$'>>} attrs
  * @param {(HTMLElement | string | null | undefined | boolean)[]} children
  * @returns {HTMLElementTagNameMap[K]}
  */
