+++ Resources/Private/JavaScript/style-mock.js
@@ -1,0 +1,12 @@
+module.exports = {
+  /**
+   * This handles (discards) style imports during Jest tests (see package.json).
+   *
+   * @returns {string}
+   */
+  process() {
+    return {
+      code: "",
+    };
+  },
+};
